+++ CloudSeed/AudioLib/Biquad.h
@@ -21,38 +21,38 @@ namespace AudioLib
 		};
 
 	private:
-		double samplerate;
-		double _gainDb;
-		double _q;
-		double a0, a1, a2, b0, b1, b2;
-		double x1, x2, y, y1, y2;
-		double gain;
+		float samplerate;
+		float _gainDb;
+		float _q;
+		float a0, a1, a2, b0, b1, b2;
+		float x1, x2, y, y1, y2;
+		float gain;
 
 	public:
 		FilterType Type;
-		double Output;
-		double Frequency;
-		double Slope;
+		float Output;
+		float Frequency;
+		float Slope;
 
 		Biquad();
-		Biquad(FilterType filterType, double samplerate);
+		Biquad(FilterType filterType, float samplerate);
 		~Biquad();
 
-		double GetSamplerate();
-		void SetSamplerate(double samplerate);
-		double GetGainDb();
-		void SetGainDb(double value);
-		double GetGain();
-		void SetGain(double value);
-		double GetQ();
-		void SetQ(double value);
-		vector<double> GetA();
-		vector<double> GetB();
+		float GetSamplerate();
+		void SetSamplerate(float samplerate);
+		float GetGainDb();
+		void SetGainDb(float value);
+		float GetGain();
+		void SetGain(float value);
+		float GetQ();
+		void SetQ(float value);
+		vector<float> GetA();
+		vector<float> GetB();
 
 		void Update();
-		double GetResponse(double freq);
+		float GetResponse(float freq);
 		
-		double inline Process(double x)
+		float inline Process(float x)
 		{
 			y = b0 * x + b1 * x1 + b2 * x2 - a1 * y1 - a2 * y2;
 			x2 = x1;
@@ -64,7 +64,7 @@ namespace AudioLib
 			return Output;
 		}
 
-		void inline Process(double* input, double* output, int len)
+		void inline Process(float* input, float* output, int len)
 		{
 			for (int i = 0; i < len; i++)
 				output[i] = Process(input[i]);
