+++ CloudSeed/AudioLib/Hp1.h
@@ -8,37 +8,37 @@ namespace AudioLib
 	class Hp1
 	{
 	private:
-		double fs;
-		double b0, a1;
-		double lpOut;
-		double cutoffHz;
+		float fs;
+		float b0, a1;
+		float lpOut;
+		float cutoffHz;
 
 	public:
-		double Output;
+		float Output;
 
 	public:
-		Hp1(double fs)
+		Hp1(float fs)
 		{
 			this->lpOut = 0;
 			this->fs = fs;
 		}
 
-		double GetSamplerate()
+		float GetSamplerate()
 		{
 			return fs;
 		}
 
-		void SetSamplerate(double samplerate)
+		void SetSamplerate(float samplerate)
 		{
 			fs = samplerate;
 		}
 
-		double GetCutoffHz()
+		float GetCutoffHz()
 		{
 			return cutoffHz;
 		}
 
-		void SetCutoffHz(double hz)
+		void SetCutoffHz(float hz)
 		{
 			cutoffHz = hz;
 			Update();
@@ -58,7 +58,7 @@ namespace AudioLib
 			b0 = 1 - alpha;
 		}
 
-		double Process(double input)
+		float Process(float input)
 		{
 			if (input == 0 && lpOut < 0.000000000001)
 			{
@@ -73,7 +73,7 @@ namespace AudioLib
 			return Output;
 		}
 
-		void Process(double* input, double* output, int len)
+		void Process(float* input, float* output, int len)
 		{
 			for (int i = 0; i < len; i++)
 				output[i] = Process(input[i]);
