+++ CloudSeed/FastSin.cpp
@@ -6,5 +6,5 @@
 
 namespace CloudSeed
 {
-	DSY_SDRAM_BSS double FastSin::data[FastSin::DataSize];
+	DSY_SDRAM_BSS float FastSin::data[FastSin::DataSize];
 }
